requirements.txt
 # 设计资讯聚合工具依赖包
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.5.0
sqlite-utils>=3.35.0
feedparser>=6.0.0
python-dateutil>=2.8.0
pytz>=2023.3
jinja2>=3.1.0
Flask>=2.3.0
Flask-CORS>=4.0.0 